Delete attachments task

+++ lib/tasks/attachments.rake
@@ -1,9 +1,12 @@
 namespace :ff do
   namespace :attachments do
-    desc "Reprocess posting attachments"
+    desc "Reprocess posting CLASS attachments for STYLES"
     task :reprocess => :environment do
-      klass  = (ENV['CLASS'] || 'Posting::Base').constantize
-      styles = (ENV['STYLES'] || '').split(',').map(&:to_sym)
+      klass  = ENV['CLASS'].constantize
+      styles = ENV['STYLES'].split(',').map { |style| style.strip.downcase.to_sym }
+      raise "CLASS needs to be defined" if klass.blank?
+      raise "STYLES needs to be defined" if styles.blank?      
+      raise "Can't reprocess original" if styles.include?('original')      
       klass.all.each do |posting|
         if posting.respond_to?(:image)
             print "#{posting.id} "
@@ -30,12 +33,18 @@ namespace :ff do
       end
     end
 
-    desc "Delete all (except the original) posting image attachments"
+    desc "Delete posting CLASS attachments for STYLE"
     task :delete => :environment do
       require 'fileutils'
-      styles = ENV['STYLES'].present? ? "{#{ENV['STYLES']}}" : '*'
-      dirs = (Dir["public/system/images/*/#{styles}"] - Dir['public/system/images/*/original'])
-      dirs.each { |dir| FileUtils.rm_rf dir }
+      klass  = ENV['CLASS'].constantize      
+      styles = ENV['STYLES'].split(',').map { |style| style.strip.downcase.to_sym }
+      raise "CLASS needs to be defined" if klass.blank?      
+      raise "STYLES needs to be defined" if styles.blank?
+      raise "Can't delete original" if styles.include?(:original)
+      klass.find_each do |posting|
+        dirs = Dir[File.join('public', 'system', 'images', posting[:id].to_s, "{#{styles.join(',')}}")] 
+        dirs.each { |dir| FileUtils.rm_rf dir }
+      end
     end
   end
 end
